Resolve issues after merge
