+++ src/argilla/client/feedback/mixin.py
@@ -13,6 +13,7 @@
 #  limitations under the License.
 
 import logging
+import warnings
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Type, Union
 from uuid import UUID
 
@@ -279,6 +280,20 @@ class ArgillaDatasetMixin:
                         disable=show_progress,
                     ):
                         for record in updated_records[i : i + PUSHING_BATCH_SIZE]:
+                            if not record.id:
+                                warnings.warn(
+                                    "A record is already pushed to Argilla but no record id found, which means"
+                                    " that the `FeedbackRecord` has been pushed to  Argilla, but hasn't been fetched,"
+                                    " so the `id` is missing. Record update will be skipped for this record"
+                                )
+                                if record.suggestions:
+                                    warnings.warn(
+                                        "The `suggestions` have been provided, but the `id`"
+                                        " is not set. To solve that, you can simply call"
+                                        " `FeedbackDataset.fetch_records()` to fetch them and"
+                                        " automatically set the `id`, to call `set_suggestions` on top of that."
+                                    )
+                                continue
                             for suggestion in record.suggestions:
                                 suggestion.question_id = question_name2id[suggestion.question_name]
                                 datasets_api_v1.set_suggestion(
